removes Recommender class; all applicable methods from that class were moved to Introspector class
